Fix mistaken generation of a build error for case-sensitive file systems.
On a case sensitive file system if more than one paths with similar names
exists, it could cause a false build error to be generated.  For example,
if both an x64 and an X64 directory existed, then it was possible that an
error would be generated saying that the file case was incorrect.

git-svn-id: https://edk2-buildtools.svn.sourceforge.net/svnroot/edk2-buildtools/trunk/BaseTools@1713 c2973edb-eda0-4c78-bc6a-9341b269661f