Fix incremental build issue for FV Extended Headers.  If only the contents of the FV Extended Header has been modified, then the change would not be reflected in the final FV/FD image.
This fix updates GenFds to update the time/date stamp on an FV INF file if the contents if the FV Extended Header is present and those contents have changed since the last time the FV Extended Header was generated.

Do not update time/date stamp if FV INF file does not exist.



git-svn-id: https://edk2-buildtools.svn.sourceforge.net/svnroot/edk2-buildtools/trunk/BaseTools@1752 c2973edb-eda0-4c78-bc6a-9341b269661f