GCC44 X64: Fix UEFI image incompatibility with addresses above 2GB
As documented in the GCC manual under the 'Intel 386 and AMD x86-64
Options' section:
http://gcc.gnu.org/onlinedocs/gcc-4.4.5/gcc/i386-and-x86_002d64-Options.html

The default setting for -mcmodel is small.  With this setting, code
is generated with the assumption that it will run in the lower 2GB
of the address space.

For UEFI compatibility, images should be relocatable to any address.
Therefore, we enable the -mcmodel=large setting.

Unfortunately with GCC 4.4, a build error occurs if both
-mcmodel=large and -Os are used.  (-Os: optimize for image size.)
Therefore, we must remove the -Os parameter, which does increase
the image size considerably.

git-svn-id: https://edk2-buildtools.svn.sourceforge.net/svnroot/edk2-buildtools/trunk/BaseTools@2069 c2973edb-eda0-4c78-bc6a-9341b269661f